# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Application'","description":"'Application for Spring Boot'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'Application for Spring Boot'"}],"paths":{"/guess-number/users":{"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/guess-number/users/{userId}/matches":{"post":{"tags":["user-controller"],"operationId":"createUserMatch","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserMatchDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MatchDto"}}}}}}},"/guess-number/users/{userId}/matches/{matchId}":{"post":{"tags":["user-controller"],"operationId":"playUserMatch","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"matchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlayUserMatchDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoundMatchDto"}}}}}}}},"components":{"schemas":{"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"username":{"type":"string"}}},"CreateUserMatchDto":{"type":"object","properties":{"difficulty":{"type":"string","enum":["EASY","MEDIUM","HARD"]}}},"MatchDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"matchDifficulty":{"type":"string","enum":["EASY","MEDIUM","HARD"]},"remainingTries":{"type":"integer","format":"int32"}}},"PlayUserMatchDto":{"type":"object","properties":{"number":{"type":"integer","format":"int32"}}},"RoundMatchDto":{"type":"object","properties":{"matchDto":{"$ref":"#/components/schemas/MatchDto"},"response":{"type":"string"}}}}}}